+++ R/spec-version.R
@@ -21,12 +21,12 @@
 vw_spec_version <- function(spec) {
 
   spec <- as_vegaspec(spec)
-  version <- .schema_type(spec[["$schema"]])
+  version <- parse_schema(spec[["$schema"]])
 
   version
 }
 
-.schema_type <- function(schema) {
+parse_schema <- function(schema) {
 
   result <- list(library = "", version = "")
 
@@ -94,4 +94,15 @@ vega_schema <- function(library = c("vega_lite", "vega"), version = NULL,
   schema <- as.character(schema)
 
   schema
+}
+
+# internal function to help test different schema versions
+with_schema <- function(value, spec) {
+  schema <- spec[["$schema"]]
+  schema_new <- sub("v(\\d+)\\.json$", glue::glue("v{value}.json"), schema)
+
+  spec_new <- spec
+  spec_new[["$schema"]] <- schema_new
+
+  spec_new
 }