+++ R/utils-null.R
@@ -1,1 +1,2 @@
-`%||%` <- rlang::`%||%`
+#' @importFrom rlang `%||%`
+NULL